refactor: 3단계 - 경로 조회 기능 (PR 피드백 2 적용)
- 테스트 코드 double 형으로 변경

+++ src/test/java/nextstep/subway/acceptance/PathServiceAcceptanceTest.java
@@ -69,7 +69,7 @@ public class PathServiceAcceptanceTest extends AcceptanceTest {
 
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
         assertThat(response.jsonPath().getList("stations.id", Long.class)).containsExactly(교대역, 남부터미널역, 양재역);
-        assertThat(response.jsonPath().getInt("distance")).isEqualTo(5);
+        assertThat(response.jsonPath().getDouble("distance")).isEqualTo(5);
     }
 
     /**
